third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_word_11.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'word'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} word`
  );
}

function makeSample(offset){
  const kSample = 'Lorem <span style="direction: ltr">ipsum dolor<div > just a test</div> sit</span> amet';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right word
testIt('ltr', 0, 'right', '^Lorem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 1, 'right', 'L^orem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 2, 'right', 'Lo^rem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 3, 'right', 'Lor^em| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 4, 'right', 'Lore^m| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 5, 'right', 'Lorem^ <span style="direction: ltr">ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 6, 'right', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 7, 'right', 'Lorem <^span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('ltr', 8, 'right', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('ltr', 9, 'right', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('ltr', 10, 'right', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('ltr', 11, 'right', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('ltr', 12, 'right', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 13, 'right', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 14, 'right', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 15, 'right', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 16, 'right', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 17, 'right', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 18, 'right', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 19, 'right', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 20, 'right', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 21, 'right', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 22, 'right', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 23, 'right', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 24, 'right', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 25, 'right', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 26, 'right', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 27, 'right', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 28, 'right', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 29, 'right', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 30, 'right', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 31, 'right', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 32, 'right', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 33, 'right', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 34, 'right', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 35, 'right', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 36, 'right', 'Lorem <span style="direction: ltr">i^psum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 37, 'right', 'Lorem <span style="direction: ltr">ip^sum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 38, 'right', 'Lorem <span style="direction: ltr">ips^um| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 39, 'right', 'Lorem <span style="direction: ltr">ipsu^m| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 40, 'right', 'Lorem <span style="direction: ltr">ipsum^ dolor|<div> just a test</div> sit</span> amet');
testIt('ltr', 41, 'right', 'Lorem <span style="direction: ltr">ipsum ^dolor|<div> just a test</div> sit</span> amet');
testIt('ltr', 42, 'right', 'Lorem <span style="direction: ltr">ipsum d^olor|<div> just a test</div> sit</span> amet');
testIt('ltr', 43, 'right', 'Lorem <span style="direction: ltr">ipsum do^lor|<div> just a test</div> sit</span> amet');
testIt('ltr', 44, 'right', 'Lorem <span style="direction: ltr">ipsum dol^or|<div> just a test</div> sit</span> amet');
testIt('ltr', 45, 'right', 'Lorem <span style="direction: ltr">ipsum dolo^r|<div> just a test</div> sit</span> amet');
testIt('ltr', 46, 'right', 'Lorem <span style="direction: ltr">ipsum dolor^<div> just| a test</div> sit</span> amet');
testIt('ltr', 47, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<^div| > just a test sit</span> amet');
testIt('ltr', 48, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('ltr', 49, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('ltr', 50, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('ltr', 51, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('ltr', 52, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('ltr', 53, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('ltr', 54, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> j^ust| a test</div> sit</span> amet');
testIt('ltr', 55, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> ju^st| a test</div> sit</span> amet');
testIt('ltr', 56, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> jus^t| a test</div> sit</span> amet');
testIt('ltr', 57, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just^ a| test</div> sit</span> amet');
testIt('ltr', 58, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just ^a| test</div> sit</span> amet');
testIt('ltr', 59, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a^ test|</div> sit</span> amet');
testIt('ltr', 60, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a ^test|</div> sit</span> amet');
testIt('ltr', 61, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a t^est|</div> sit</span> amet');
testIt('ltr', 62, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a te^st|</div> sit</span> amet');
testIt('ltr', 63, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a tes^t|</div> sit</span> amet');
testIt('ltr', 64, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test^</div> sit|</span> amet');
testIt('ltr', 65, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test<^/div|> sit amet</div></span>');
testIt('ltr', 66, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a testdiv^ sit| amet</div></span>');
testIt('ltr', 67, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 68, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 69, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 70, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('ltr', 71, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('ltr', 72, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> s^it|</span> amet');
testIt('ltr', 73, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> si^t|</span> amet');
testIt('ltr', 74, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('ltr', 75, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit<^/span|> amet</span>');
testIt('ltr', 76, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sitspan^ amet|</span>');
testIt('ltr', 77, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 78, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 79, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 80, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 81, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('ltr', 82, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ^amet|');
testIt('ltr', 83, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> a^met|');
testIt('ltr', 84, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> am^et|');
testIt('ltr', 85, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ame^t|');

// ltr left word
testIt('ltr', 0, 'left', '|Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 1, 'left', '|L^orem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 2, 'left', '|Lo^rem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 3, 'left', '|Lor^em <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 4, 'left', '|Lore^m <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 5, 'left', '|Lorem^ <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 6, 'left', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 7, 'left', '|Lorem <^span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('ltr', 8, 'left', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('ltr', 9, 'left', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('ltr', 10, 'left', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('ltr', 11, 'left', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('ltr', 12, 'left', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 13, 'left', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 14, 'left', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 15, 'left', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 16, 'left', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 17, 'left', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 18, 'left', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 19, 'left', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 20, 'left', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 21, 'left', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 22, 'left', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 23, 'left', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 24, 'left', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 25, 'left', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 26, 'left', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 27, 'left', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 28, 'left', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 29, 'left', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 30, 'left', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 31, 'left', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 32, 'left', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 33, 'left', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 34, 'left', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 35, 'left', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 36, 'left', 'Lorem <span style="direction: ltr">|i^psum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 37, 'left', 'Lorem <span style="direction: ltr">|ip^sum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 38, 'left', 'Lorem <span style="direction: ltr">|ips^um dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 39, 'left', 'Lorem <span style="direction: ltr">|ipsu^m dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 40, 'left', 'Lorem <span style="direction: ltr">|ipsum^ dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 41, 'left', 'Lorem <span style="direction: ltr">|ipsum ^dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 42, 'left', 'Lorem <span style="direction: ltr">ipsum |d^olor<div> just a test</div> sit</span> amet');
testIt('ltr', 43, 'left', 'Lorem <span style="direction: ltr">ipsum |do^lor<div> just a test</div> sit</span> amet');
testIt('ltr', 44, 'left', 'Lorem <span style="direction: ltr">ipsum |dol^or<div> just a test</div> sit</span> amet');
testIt('ltr', 45, 'left', 'Lorem <span style="direction: ltr">ipsum |dolo^r<div> just a test</div> sit</span> amet');
testIt('ltr', 46, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor^<div> just a test</div> sit</span> amet');
testIt('ltr', 47, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor<^div > just a test sit</span> amet');
testIt('ltr', 48, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('ltr', 49, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('ltr', 50, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('ltr', 51, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('ltr', 52, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('ltr', 53, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('ltr', 54, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |j^ust a test</div> sit</span> amet');
testIt('ltr', 55, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |ju^st a test</div> sit</span> amet');
testIt('ltr', 56, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |jus^t a test</div> sit</span> amet');
testIt('ltr', 57, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just^ a test</div> sit</span> amet');
testIt('ltr', 58, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just ^a test</div> sit</span> amet');
testIt('ltr', 59, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a^ test</div> sit</span> amet');
testIt('ltr', 60, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a ^test</div> sit</span> amet');
testIt('ltr', 61, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |t^est</div> sit</span> amet');
testIt('ltr', 62, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |te^st</div> sit</span> amet');
testIt('ltr', 63, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |tes^t</div> sit</span> amet');
testIt('ltr', 64, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^</div> sit</span> amet');
testIt('ltr', 65, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test<^/div> sit amet</div></span>');
testIt('ltr', 66, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^div sit amet</div></span>');
testIt('ltr', 67, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 68, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 69, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 70, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('ltr', 71, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('ltr', 72, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |s^it</span> amet');
testIt('ltr', 73, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |si^t</span> amet');
testIt('ltr', 74, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('ltr', 75, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit<^/span> amet</span>');
testIt('ltr', 76, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^span amet</span>');
testIt('ltr', 77, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 78, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 79, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 80, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 81, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('ltr', 82, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit</span> ^amet');
testIt('ltr', 83, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |a^met');
testIt('ltr', 84, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |am^et');
testIt('ltr', 85, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |ame^t');

// ltr forward word
testIt('ltr', 0, 'forward', '^Lorem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 1, 'forward', 'L^orem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 2, 'forward', 'Lo^rem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 3, 'forward', 'Lor^em| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 4, 'forward', 'Lore^m| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 5, 'forward', 'Lorem^ <span style="direction: ltr">ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 6, 'forward', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 7, 'forward', 'Lorem <^span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('ltr', 8, 'forward', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('ltr', 9, 'forward', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('ltr', 10, 'forward', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('ltr', 11, 'forward', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('ltr', 12, 'forward', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 13, 'forward', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 14, 'forward', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 15, 'forward', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 16, 'forward', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 17, 'forward', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 18, 'forward', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 19, 'forward', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 20, 'forward', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 21, 'forward', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 22, 'forward', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 23, 'forward', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 24, 'forward', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 25, 'forward', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 26, 'forward', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 27, 'forward', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 28, 'forward', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 29, 'forward', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 30, 'forward', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 31, 'forward', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 32, 'forward', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 33, 'forward', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 34, 'forward', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 35, 'forward', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 36, 'forward', 'Lorem <span style="direction: ltr">i^psum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 37, 'forward', 'Lorem <span style="direction: ltr">ip^sum| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 38, 'forward', 'Lorem <span style="direction: ltr">ips^um| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 39, 'forward', 'Lorem <span style="direction: ltr">ipsu^m| dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 40, 'forward', 'Lorem <span style="direction: ltr">ipsum^ dolor|<div> just a test</div> sit</span> amet');
testIt('ltr', 41, 'forward', 'Lorem <span style="direction: ltr">ipsum ^dolor|<div> just a test</div> sit</span> amet');
testIt('ltr', 42, 'forward', 'Lorem <span style="direction: ltr">ipsum d^olor|<div> just a test</div> sit</span> amet');
testIt('ltr', 43, 'forward', 'Lorem <span style="direction: ltr">ipsum do^lor|<div> just a test</div> sit</span> amet');
testIt('ltr', 44, 'forward', 'Lorem <span style="direction: ltr">ipsum dol^or|<div> just a test</div> sit</span> amet');
testIt('ltr', 45, 'forward', 'Lorem <span style="direction: ltr">ipsum dolo^r|<div> just a test</div> sit</span> amet');
testIt('ltr', 46, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor^<div> just| a test</div> sit</span> amet');
testIt('ltr', 47, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<^div| > just a test sit</span> amet');
testIt('ltr', 48, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('ltr', 49, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('ltr', 50, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('ltr', 51, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('ltr', 52, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('ltr', 53, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('ltr', 54, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> j^ust| a test</div> sit</span> amet');
testIt('ltr', 55, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> ju^st| a test</div> sit</span> amet');
testIt('ltr', 56, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> jus^t| a test</div> sit</span> amet');
testIt('ltr', 57, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just^ a| test</div> sit</span> amet');
testIt('ltr', 58, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just ^a| test</div> sit</span> amet');
testIt('ltr', 59, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a^ test|</div> sit</span> amet');
testIt('ltr', 60, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a ^test|</div> sit</span> amet');
testIt('ltr', 61, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a t^est|</div> sit</span> amet');
testIt('ltr', 62, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a te^st|</div> sit</span> amet');
testIt('ltr', 63, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a tes^t|</div> sit</span> amet');
testIt('ltr', 64, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test^</div> sit|</span> amet');
testIt('ltr', 65, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test<^/div|> sit amet</div></span>');
testIt('ltr', 66, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a testdiv^ sit| amet</div></span>');
testIt('ltr', 67, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 68, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 69, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 70, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('ltr', 71, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('ltr', 72, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> s^it|</span> amet');
testIt('ltr', 73, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> si^t|</span> amet');
testIt('ltr', 74, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('ltr', 75, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit<^/span|> amet</span>');
testIt('ltr', 76, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sitspan^ amet|</span>');
testIt('ltr', 77, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 78, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 79, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 80, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 81, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('ltr', 82, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ^amet|');
testIt('ltr', 83, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> a^met|');
testIt('ltr', 84, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> am^et|');
testIt('ltr', 85, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ame^t|');

// ltr backward word
testIt('ltr', 0, 'backward', '|Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 1, 'backward', '|L^orem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 2, 'backward', '|Lo^rem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 3, 'backward', '|Lor^em <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 4, 'backward', '|Lore^m <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 5, 'backward', '|Lorem^ <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 6, 'backward', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 7, 'backward', '|Lorem <^span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('ltr', 8, 'backward', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('ltr', 9, 'backward', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('ltr', 10, 'backward', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('ltr', 11, 'backward', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('ltr', 12, 'backward', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 13, 'backward', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 14, 'backward', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 15, 'backward', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 16, 'backward', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 17, 'backward', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 18, 'backward', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 19, 'backward', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 20, 'backward', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 21, 'backward', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 22, 'backward', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 23, 'backward', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 24, 'backward', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 25, 'backward', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 26, 'backward', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 27, 'backward', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 28, 'backward', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 29, 'backward', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 30, 'backward', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 31, 'backward', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 32, 'backward', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 33, 'backward', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 34, 'backward', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 35, 'backward', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 36, 'backward', 'Lorem <span style="direction: ltr">|i^psum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 37, 'backward', 'Lorem <span style="direction: ltr">|ip^sum dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 38, 'backward', 'Lorem <span style="direction: ltr">|ips^um dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 39, 'backward', 'Lorem <span style="direction: ltr">|ipsu^m dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 40, 'backward', 'Lorem <span style="direction: ltr">|ipsum^ dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 41, 'backward', 'Lorem <span style="direction: ltr">|ipsum ^dolor<div> just a test</div> sit</span> amet');
testIt('ltr', 42, 'backward', 'Lorem <span style="direction: ltr">ipsum |d^olor<div> just a test</div> sit</span> amet');
testIt('ltr', 43, 'backward', 'Lorem <span style="direction: ltr">ipsum |do^lor<div> just a test</div> sit</span> amet');
testIt('ltr', 44, 'backward', 'Lorem <span style="direction: ltr">ipsum |dol^or<div> just a test</div> sit</span> amet');
testIt('ltr', 45, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolo^r<div> just a test</div> sit</span> amet');
testIt('ltr', 46, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor^<div> just a test</div> sit</span> amet');
testIt('ltr', 47, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor<^div > just a test sit</span> amet');
testIt('ltr', 48, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('ltr', 49, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('ltr', 50, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('ltr', 51, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('ltr', 52, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('ltr', 53, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('ltr', 54, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |j^ust a test</div> sit</span> amet');
testIt('ltr', 55, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |ju^st a test</div> sit</span> amet');
testIt('ltr', 56, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |jus^t a test</div> sit</span> amet');
testIt('ltr', 57, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just^ a test</div> sit</span> amet');
testIt('ltr', 58, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just ^a test</div> sit</span> amet');
testIt('ltr', 59, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a^ test</div> sit</span> amet');
testIt('ltr', 60, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a ^test</div> sit</span> amet');
testIt('ltr', 61, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |t^est</div> sit</span> amet');
testIt('ltr', 62, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |te^st</div> sit</span> amet');
testIt('ltr', 63, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |tes^t</div> sit</span> amet');
testIt('ltr', 64, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^</div> sit</span> amet');
testIt('ltr', 65, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test<^/div> sit amet</div></span>');
testIt('ltr', 66, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^div sit amet</div></span>');
testIt('ltr', 67, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 68, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 69, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('ltr', 70, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('ltr', 71, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('ltr', 72, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |s^it</span> amet');
testIt('ltr', 73, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |si^t</span> amet');
testIt('ltr', 74, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('ltr', 75, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit<^/span> amet</span>');
testIt('ltr', 76, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^span amet</span>');
testIt('ltr', 77, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 78, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 79, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 80, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('ltr', 81, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('ltr', 82, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit</span> ^amet');
testIt('ltr', 83, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |a^met');
testIt('ltr', 84, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |am^et');
testIt('ltr', 85, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |ame^t');

// rtl right word
testIt('rtl', 0, 'right', '^Lorem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 1, 'right', 'L^orem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 2, 'right', 'Lo^rem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 3, 'right', 'Lor^em| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 4, 'right', 'Lore^m| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 5, 'right', 'Lorem^ <span style="direction: ltr">ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 6, 'right', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 7, 'right', 'Lorem <^span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('rtl', 8, 'right', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('rtl', 9, 'right', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('rtl', 10, 'right', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('rtl', 11, 'right', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('rtl', 12, 'right', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 13, 'right', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 14, 'right', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 15, 'right', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 16, 'right', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 17, 'right', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 18, 'right', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 19, 'right', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 20, 'right', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 21, 'right', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 22, 'right', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 23, 'right', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 24, 'right', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 25, 'right', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 26, 'right', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 27, 'right', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 28, 'right', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 29, 'right', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 30, 'right', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 31, 'right', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 32, 'right', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 33, 'right', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 34, 'right', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 35, 'right', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 36, 'right', 'Lorem <span style="direction: ltr">i^psum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 37, 'right', 'Lorem <span style="direction: ltr">ip^sum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 38, 'right', 'Lorem <span style="direction: ltr">ips^um| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 39, 'right', 'Lorem <span style="direction: ltr">ipsu^m| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 40, 'right', 'Lorem <span style="direction: ltr">ipsum^ dolor|<div> just a test</div> sit</span> amet');
testIt('rtl', 41, 'right', 'Lorem <span style="direction: ltr">ipsum ^dolor|<div> just a test</div> sit</span> amet');
testIt('rtl', 42, 'right', 'Lorem <span style="direction: ltr">ipsum d^olor|<div> just a test</div> sit</span> amet');
testIt('rtl', 43, 'right', 'Lorem <span style="direction: ltr">ipsum do^lor|<div> just a test</div> sit</span> amet');
testIt('rtl', 44, 'right', 'Lorem <span style="direction: ltr">ipsum dol^or|<div> just a test</div> sit</span> amet');
testIt('rtl', 45, 'right', 'Lorem <span style="direction: ltr">ipsum dolo^r|<div> just a test</div> sit</span> amet');
testIt('rtl', 46, 'right', 'Lorem <span style="direction: ltr">ipsum dolor^<div> just| a test</div> sit</span> amet');
testIt('rtl', 47, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<^div| > just a test sit</span> amet');
testIt('rtl', 48, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('rtl', 49, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('rtl', 50, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('rtl', 51, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('rtl', 52, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('rtl', 53, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('rtl', 54, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> j^ust| a test</div> sit</span> amet');
testIt('rtl', 55, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> ju^st| a test</div> sit</span> amet');
testIt('rtl', 56, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> jus^t| a test</div> sit</span> amet');
testIt('rtl', 57, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just^ a| test</div> sit</span> amet');
testIt('rtl', 58, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just ^a| test</div> sit</span> amet');
testIt('rtl', 59, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a^ test|</div> sit</span> amet');
testIt('rtl', 60, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a ^test|</div> sit</span> amet');
testIt('rtl', 61, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a t^est|</div> sit</span> amet');
testIt('rtl', 62, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a te^st|</div> sit</span> amet');
testIt('rtl', 63, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a tes^t|</div> sit</span> amet');
testIt('rtl', 64, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test^</div> sit|</span> amet');
testIt('rtl', 65, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test<^/div|> sit amet</div></span>');
testIt('rtl', 66, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a testdiv^ sit| amet</div></span>');
testIt('rtl', 67, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 68, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 69, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 70, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('rtl', 71, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('rtl', 72, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> s^it|</span> amet');
testIt('rtl', 73, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> si^t|</span> amet');
testIt('rtl', 74, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('rtl', 75, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit<^/span|> amet</span>');
testIt('rtl', 76, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sitspan^ amet|</span>');
testIt('rtl', 77, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 78, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 79, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 80, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 81, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('rtl', 82, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ^amet|');
testIt('rtl', 83, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> a^met|');
testIt('rtl', 84, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> am^et|');
testIt('rtl', 85, 'right', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ame^t|');

// rtl left word
testIt('rtl', 0, 'left', '|Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 1, 'left', '|L^orem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 2, 'left', '|Lo^rem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 3, 'left', '|Lor^em <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 4, 'left', '|Lore^m <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 5, 'left', '|Lorem^ <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 6, 'left', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 7, 'left', '|Lorem <^span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('rtl', 8, 'left', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('rtl', 9, 'left', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('rtl', 10, 'left', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('rtl', 11, 'left', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('rtl', 12, 'left', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 13, 'left', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 14, 'left', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 15, 'left', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 16, 'left', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 17, 'left', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 18, 'left', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 19, 'left', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 20, 'left', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 21, 'left', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 22, 'left', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 23, 'left', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 24, 'left', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 25, 'left', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 26, 'left', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 27, 'left', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 28, 'left', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 29, 'left', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 30, 'left', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 31, 'left', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 32, 'left', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 33, 'left', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 34, 'left', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 35, 'left', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 36, 'left', 'Lorem <span style="direction: ltr">|i^psum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 37, 'left', 'Lorem <span style="direction: ltr">|ip^sum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 38, 'left', 'Lorem <span style="direction: ltr">|ips^um dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 39, 'left', 'Lorem <span style="direction: ltr">|ipsu^m dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 40, 'left', 'Lorem <span style="direction: ltr">|ipsum^ dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 41, 'left', 'Lorem <span style="direction: ltr">|ipsum ^dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 42, 'left', 'Lorem <span style="direction: ltr">ipsum |d^olor<div> just a test</div> sit</span> amet');
testIt('rtl', 43, 'left', 'Lorem <span style="direction: ltr">ipsum |do^lor<div> just a test</div> sit</span> amet');
testIt('rtl', 44, 'left', 'Lorem <span style="direction: ltr">ipsum |dol^or<div> just a test</div> sit</span> amet');
testIt('rtl', 45, 'left', 'Lorem <span style="direction: ltr">ipsum |dolo^r<div> just a test</div> sit</span> amet');
testIt('rtl', 46, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor^<div> just a test</div> sit</span> amet');
testIt('rtl', 47, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor<^div > just a test sit</span> amet');
testIt('rtl', 48, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('rtl', 49, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('rtl', 50, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('rtl', 51, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('rtl', 52, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('rtl', 53, 'left', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('rtl', 54, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |j^ust a test</div> sit</span> amet');
testIt('rtl', 55, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |ju^st a test</div> sit</span> amet');
testIt('rtl', 56, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |jus^t a test</div> sit</span> amet');
testIt('rtl', 57, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just^ a test</div> sit</span> amet');
testIt('rtl', 58, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just ^a test</div> sit</span> amet');
testIt('rtl', 59, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a^ test</div> sit</span> amet');
testIt('rtl', 60, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a ^test</div> sit</span> amet');
testIt('rtl', 61, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |t^est</div> sit</span> amet');
testIt('rtl', 62, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |te^st</div> sit</span> amet');
testIt('rtl', 63, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |tes^t</div> sit</span> amet');
testIt('rtl', 64, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^</div> sit</span> amet');
testIt('rtl', 65, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test<^/div> sit amet</div></span>');
testIt('rtl', 66, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^div sit amet</div></span>');
testIt('rtl', 67, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 68, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 69, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 70, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('rtl', 71, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('rtl', 72, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |s^it</span> amet');
testIt('rtl', 73, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |si^t</span> amet');
testIt('rtl', 74, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('rtl', 75, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit<^/span> amet</span>');
testIt('rtl', 76, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^span amet</span>');
testIt('rtl', 77, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 78, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 79, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 80, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 81, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('rtl', 82, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit</span> ^amet');
testIt('rtl', 83, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |a^met');
testIt('rtl', 84, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |am^et');
testIt('rtl', 85, 'left', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |ame^t');

// rtl forward word
testIt('rtl', 0, 'forward', '^Lorem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 1, 'forward', 'L^orem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 2, 'forward', 'Lo^rem| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 3, 'forward', 'Lor^em| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 4, 'forward', 'Lore^m| <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 5, 'forward', 'Lorem^ <span style="direction: ltr">ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 6, 'forward', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 7, 'forward', 'Lorem <^span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('rtl', 8, 'forward', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('rtl', 9, 'forward', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('rtl', 10, 'forward', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('rtl', 11, 'forward', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('rtl', 12, 'forward', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 13, 'forward', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 14, 'forward', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 15, 'forward', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 16, 'forward', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 17, 'forward', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 18, 'forward', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 19, 'forward', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 20, 'forward', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 21, 'forward', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 22, 'forward', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 23, 'forward', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 24, 'forward', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 25, 'forward', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 26, 'forward', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 27, 'forward', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 28, 'forward', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 29, 'forward', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 30, 'forward', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 31, 'forward', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 32, 'forward', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 33, 'forward', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 34, 'forward', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 35, 'forward', 'Lorem <span style="direction: ltr">^ipsum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 36, 'forward', 'Lorem <span style="direction: ltr">i^psum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 37, 'forward', 'Lorem <span style="direction: ltr">ip^sum| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 38, 'forward', 'Lorem <span style="direction: ltr">ips^um| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 39, 'forward', 'Lorem <span style="direction: ltr">ipsu^m| dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 40, 'forward', 'Lorem <span style="direction: ltr">ipsum^ dolor|<div> just a test</div> sit</span> amet');
testIt('rtl', 41, 'forward', 'Lorem <span style="direction: ltr">ipsum ^dolor|<div> just a test</div> sit</span> amet');
testIt('rtl', 42, 'forward', 'Lorem <span style="direction: ltr">ipsum d^olor|<div> just a test</div> sit</span> amet');
testIt('rtl', 43, 'forward', 'Lorem <span style="direction: ltr">ipsum do^lor|<div> just a test</div> sit</span> amet');
testIt('rtl', 44, 'forward', 'Lorem <span style="direction: ltr">ipsum dol^or|<div> just a test</div> sit</span> amet');
testIt('rtl', 45, 'forward', 'Lorem <span style="direction: ltr">ipsum dolo^r|<div> just a test</div> sit</span> amet');
testIt('rtl', 46, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor^<div> just| a test</div> sit</span> amet');
testIt('rtl', 47, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<^div| > just a test sit</span> amet');
testIt('rtl', 48, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('rtl', 49, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('rtl', 50, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('rtl', 51, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('rtl', 52, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('rtl', 53, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> ^just| a test</div> sit</span> amet');
testIt('rtl', 54, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> j^ust| a test</div> sit</span> amet');
testIt('rtl', 55, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> ju^st| a test</div> sit</span> amet');
testIt('rtl', 56, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> jus^t| a test</div> sit</span> amet');
testIt('rtl', 57, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just^ a| test</div> sit</span> amet');
testIt('rtl', 58, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just ^a| test</div> sit</span> amet');
testIt('rtl', 59, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a^ test|</div> sit</span> amet');
testIt('rtl', 60, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a ^test|</div> sit</span> amet');
testIt('rtl', 61, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a t^est|</div> sit</span> amet');
testIt('rtl', 62, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a te^st|</div> sit</span> amet');
testIt('rtl', 63, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a tes^t|</div> sit</span> amet');
testIt('rtl', 64, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test^</div> sit|</span> amet');
testIt('rtl', 65, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test<^/div|> sit amet</div></span>');
testIt('rtl', 66, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a testdiv^ sit| amet</div></span>');
testIt('rtl', 67, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 68, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 69, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 70, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('rtl', 71, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> ^sit|</span> amet');
testIt('rtl', 72, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> s^it|</span> amet');
testIt('rtl', 73, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> si^t|</span> amet');
testIt('rtl', 74, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('rtl', 75, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit<^/span|> amet</span>');
testIt('rtl', 76, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sitspan^ amet|</span>');
testIt('rtl', 77, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 78, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 79, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 80, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 81, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span>^ amet|');
testIt('rtl', 82, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ^amet|');
testIt('rtl', 83, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> a^met|');
testIt('rtl', 84, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> am^et|');
testIt('rtl', 85, 'forward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> ame^t|');

// rtl backward word
testIt('rtl', 0, 'backward', '|Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 1, 'backward', '|L^orem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 2, 'backward', '|Lo^rem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 3, 'backward', '|Lor^em <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 4, 'backward', '|Lore^m <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 5, 'backward', '|Lorem^ <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 6, 'backward', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 7, 'backward', '|Lorem <^span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet');
testIt('rtl', 8, 'backward', 'Lorem <s|pan style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</s|pan>');
testIt('rtl', 9, 'backward', 'Lorem <sp|an style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</sp|an>');
testIt('rtl', 10, 'backward', 'Lorem <spa|n style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</spa|n>');
testIt('rtl', 11, 'backward', 'Lorem <span| style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span|>');
testIt('rtl', 12, 'backward', 'Lorem <span |style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 13, 'backward', 'Lorem <span s|tyle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 14, 'backward', 'Lorem <span st|yle="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 15, 'backward', 'Lorem <span sty|le="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 16, 'backward', 'Lorem <span styl|e="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 17, 'backward', 'Lorem <span style|="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 18, 'backward', 'Lorem <span ltr" style="|&quot;direction:">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 19, 'backward', 'Lorem <span style="|direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 20, 'backward', 'Lorem <span style="d|irection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 21, 'backward', 'Lorem <span style="di|rection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 22, 'backward', 'Lorem <span style="dir|ection: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 23, 'backward', 'Lorem <span style="dire|ction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 24, 'backward', 'Lorem <span style="direc|tion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 25, 'backward', 'Lorem <span style="direct|ion: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 26, 'backward', 'Lorem <span style="directi|on: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 27, 'backward', 'Lorem <span style="directio|n: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 28, 'backward', 'Lorem <span style="direction|: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 29, 'backward', 'Lorem <span style="direction:| ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 30, 'backward', 'Lorem <span style="direction: |ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 31, 'backward', 'Lorem <span style="direction: l|tr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 32, 'backward', 'Lorem <span style="direction: lt|r">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 33, 'backward', 'Lorem <span style="direction: ltr|">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 34, 'backward', 'Lorem <span | style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 35, 'backward', '|Lorem ^<span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 36, 'backward', 'Lorem <span style="direction: ltr">|i^psum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 37, 'backward', 'Lorem <span style="direction: ltr">|ip^sum dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 38, 'backward', 'Lorem <span style="direction: ltr">|ips^um dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 39, 'backward', 'Lorem <span style="direction: ltr">|ipsu^m dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 40, 'backward', 'Lorem <span style="direction: ltr">|ipsum^ dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 41, 'backward', 'Lorem <span style="direction: ltr">|ipsum ^dolor<div> just a test</div> sit</span> amet');
testIt('rtl', 42, 'backward', 'Lorem <span style="direction: ltr">ipsum |d^olor<div> just a test</div> sit</span> amet');
testIt('rtl', 43, 'backward', 'Lorem <span style="direction: ltr">ipsum |do^lor<div> just a test</div> sit</span> amet');
testIt('rtl', 44, 'backward', 'Lorem <span style="direction: ltr">ipsum |dol^or<div> just a test</div> sit</span> amet');
testIt('rtl', 45, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolo^r<div> just a test</div> sit</span> amet');
testIt('rtl', 46, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor^<div> just a test</div> sit</span> amet');
testIt('rtl', 47, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor<^div > just a test sit</span> amet');
testIt('rtl', 48, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<d|iv> just a test sit</d|iv></span> amet');
testIt('rtl', 49, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<di|v> just a test sit</di|v></span> amet');
testIt('rtl', 50, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div|> just a test sit</div|></span> amet');
testIt('rtl', 51, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div |> just a test</div> sit</span> amet');
testIt('rtl', 52, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('rtl', 53, 'backward', 'Lorem <span style="direction: ltr">ipsum |dolor<div>^ just a test</div> sit</span> amet');
testIt('rtl', 54, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |j^ust a test</div> sit</span> amet');
testIt('rtl', 55, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |ju^st a test</div> sit</span> amet');
testIt('rtl', 56, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |jus^t a test</div> sit</span> amet');
testIt('rtl', 57, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just^ a test</div> sit</span> amet');
testIt('rtl', 58, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> |just ^a test</div> sit</span> amet');
testIt('rtl', 59, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a^ test</div> sit</span> amet');
testIt('rtl', 60, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just |a ^test</div> sit</span> amet');
testIt('rtl', 61, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |t^est</div> sit</span> amet');
testIt('rtl', 62, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |te^st</div> sit</span> amet');
testIt('rtl', 63, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |tes^t</div> sit</span> amet');
testIt('rtl', 64, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^</div> sit</span> amet');
testIt('rtl', 65, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test<^/div> sit amet</div></span>');
testIt('rtl', 66, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test^div sit amet</div></span>');
testIt('rtl', 67, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 68, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 69, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test sit amet</div></span>');
testIt('rtl', 70, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('rtl', 71, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a |test</div>^ sit</span> amet');
testIt('rtl', 72, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |s^it</span> amet');
testIt('rtl', 73, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |si^t</span> amet');
testIt('rtl', 74, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('rtl', 75, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit<^/span> amet</span>');
testIt('rtl', 76, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^span amet</span>');
testIt('rtl', 77, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 78, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 79, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 80, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit amet</span>');
testIt('rtl', 81, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit^</span> amet');
testIt('rtl', 82, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> |sit</span> ^amet');
testIt('rtl', 83, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |a^met');
testIt('rtl', 84, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |am^et');
testIt('rtl', 85, 'backward', 'Lorem <span style="direction: ltr">ipsum dolor<div> just a test</div> sit</span> |ame^t');
</script>
